Started Stage 1 of development
On the branch stage 1, this will be the branch where stage 1 of the
development will take place. Stage 1 involves creating the basic
gameplay including: basic test movement of player, basic obstacles,
and a simulation of endless scolling using a background.

This commit is the beginning of this stage of development.